Make Cpp2Worker::getMetadataForService const
Summary:
This function does not modify any observable state of `Cpp2Worker`. I don't count the state in `PerServiceMetadata` as part of `Cpp2Worker`'s state. Only the pointers to `PerServiceMetadata` are, and they are not mutated in this function.

We need the map to be mutable because the `PerServiceMetadata` object returned should be mutable.

Reviewed By: andriigrynenko

Differential Revision: D29243628

fbshipit-source-id: 9ddf1e5dde24ba0c019fd5df083562b3a328c0ec