Simplify RocketServerConnection::SocketDrainer
Summary: Existing SocketDrainer logic is quite fragile and still results in some races (client seeing connection being reset before getting the final message). This diff simplfies the SocketDrainer to unconditionally wait until timeout expires (or client resets the connection) before closing the connection. The only downside of this approach is that it may take longer for the server to close a connection if client is not cooperating (doesn't close the connection after receiving the close/drain event) or if client deteched the channel from the event base.

Reviewed By: iahs

Differential Revision: D29222677

fbshipit-source-id: 7351aeb5d7c2031247060d46069f4031564601d7