Let ThriftProcessor directly access ThriftServer
Summary:
Thanks to circular dependency squashing (both includes and buck), we can now directly access the server object instead of jumping through hoops.

Note that we're preserving the existing behavior of creating only one `AsyncProcessor` for the lifetime of the server. Making this per-connection (like the header and rocket) requires some additional refactoring. Afterwards, we can remove passing `ThriftServer` in the constructor

Reviewed By: andriigrynenko

Differential Revision: D29219363

fbshipit-source-id: a18adb868fd063173cb8ab2363c6b613c93492af