Remove getStats() API from ThreadManager
Summary:
ThreadManager is exposing `getStats()` API. And the data it computes  is not used anywhere.
To add insult to the injury , ThriftSever was enabling the collection of these stats by default,  and there is a spin lock involved to collect the stats (see reportTaskStats implementation) which can cause significant CPU costs.  The `getStats` API computes average running and execution time for tasks. But same functionality is currently available using the TM observer as well. Due to these reason, we should reddiff the functionality.

Reviewed By: yfeldblum

Differential Revision: D28488094

fbshipit-source-id: f53a73249b978c993884e6c3802b5d053d3de6fe