remove pipeline channel
Summary: Pipeline channel has fundamental issues around how to deal with restarting connections and error isolation within connections. While we could make a connection pool, these largely are dependent on how we connect, and is already solved in SR. As such, I'm proposing we just remove this and depend on higher-layer constructs to solve this.

Reviewed By: dlespiau

Differential Revision: D29033319

fbshipit-source-id: 0baedf2d499c9d08e486287d996fa9db8e7895c7