Update const values within lenient const types to their specific type
Summary:
Thrift allows some leniency in type declarations (eg you can declare `const bool x = 1`), and these go in the AST as a const with a type of bool with a const value with type int, this corrects it so that it's a bool, and does the same for float/double.

Also adds mutation checks on structured annotations.

Reviewed By: yfeldblum

Differential Revision: D29153617

fbshipit-source-id: 737100db434b15e6f12792e7c864f21e5aece1e4