Egress memory limit
Summary:
The egress limit is a dynamic configuration parameter that allows services to
set a limit on the memory available for inflight responses, meaning responses
that are queued on the server pending delivery to clients. Response queues can
lead to OOM crashes if clients are unable to process responses fast enough. A
typical example is a service that accepts small requests but generates large
processing-heavy responses which may cause clients to fall behind.

Reviewed By: praihan

Differential Revision: D29239832

fbshipit-source-id: dfb2777bd356f9aa1634b84e03758be829612a6c