bugfix: non-optional lazy field isn't serialized when terse writes is enabled
Summary:
Currently when terse writes is enabled, for non-optional lazy field, we check whether data member equals default value. If field is lazy and we did not deserialize it, we will skip this field.

After this diff,

1. If lazy field is un-deserialized, we always serialize it (i.e. bypass terse write logic). Note that this won't happen if user creates lazy thrift structure and fill the fields manually. This will only happen if the lazy thrift is deserialized from another upstream.
2. If lazy field is not un-deserialized, we still check whether the field equals default value (i.e. still use terse write logic here)

Optional field doesn't have this problem -- we always serialize as long as field is not empty

This diff also added terse writes to existing unit-test.

Reviewed By: luciang

Differential Revision: D29467379

fbshipit-source-id: ceff033d8f30f7c32718228850b83c54a0445355