Java: Properly close interaction with metadatapush
Summary:
Add proper closing of interaction when the interaction object is closed of
disposed of.
Indeed, each interaction object has either a `close` method (for Sync
and Async clients) or a `dispose` method (for Reactive clients).

This sends a message to the server indicating the interaction
can be closed and server-side resources can be released.

Reviewed By: avalonalex

Differential Revision: D29406751

fbshipit-source-id: b18c6c4c18ca3708ec1fee8f117e3943e0b55230