Change ResponseChannelRequest stream/sink API to consume request object
Summary: There is a tricky requirement not to destroy the request object inline in streaming callbacks. To protect callsites from forgetting to reschedule the deletion we have the API consume the pointer and reschedule internally. This only affects the callback overloads, which are typically used by proxies (where this mistake is easy to make).

Reviewed By: andriigrynenko

Differential Revision: D29565434

fbshipit-source-id: 5bd6d582ea29aac61f6829e4409a194849365c2f