fix inconsistency between lazy and non-lazy field
Summary:
When deserializing lazy field, if the underlying IOBuf already has value, we won't deserialize it. Instead, we will overwrite it.

This introduces a subtle inconsistency between lazy and non-lazy field:

If a field is structure and it already has value, and the field itself exists in serialized data, but in the inner structure, there is no fields in serialized data.

In this case, even though field in serialized data is empty, for non-lazy field, we will keep the field untouched since when deserializing structure, it won't touch field that's not in serialized data.

However, currently since we replace IOBuf completely, for lazy field, this field will be empty.

This diff fixed the inconsistency.

Reviewed By: yfeldblum

Differential Revision: D29579214

fbshipit-source-id: 7392a6e92075fcec8215cda984c4ee03f961e3e8