fix inconsistency between lazy and non-lazy field
Summary:
Previously we introduced a new optimization that when serializing thrift structure, for lazy field, we copy un-deserialized iobuf to output without deserialize it, then reserialize it.

This will actually change the behavior between lazy and non-lazy field in a subtle way.

If the lazy field itself is structure that already has data, and it also has un-deserialized iobuf. The result will be different if we pass iobuf and ignore existing data vs deserialize iobuf on existing data (since if inner field isn't in serialized data, we will keep that field untouched)

On the other hand, if field itself isn't structure (e.g. it's container, primitive type or even cpp.ref type), this is fine since in this case during deserialization, we will always clear any existing data first.

Reviewed By: yfeldblum, luciang

Differential Revision: D29600240

fbshipit-source-id: d6db9356d3ed58dee68895b99b41b19f5cd35b10