Add test for cpp.type + cpp.adapter on a typedef.
Summary:
And update test cases to use the new APIs.

To demonstrate the bug where:
  typedef binary (cpp.type="folly::IOBuf", cpp.adapter="MyAdapter2") AdaptedTypeTypeDef;

correctly defines the typedef as:
  typedef ::apache::thrift::adapt_detail::adapted_t<MyAdapter1, folly::IOBuf> AdaptedTypeTypeDef;

but then also uses the typedef name for the "standard type" (see `type_resolver_test.cc:413`). The standard type is the type that the runtime is expecting, so should not be the adapted type. It should actually just be `folly::IOBuf` in this case.

Reviewed By: yfeldblum

Differential Revision: D29642549

fbshipit-source-id: 56c94b88330bde8c6ff211ef40051b6bd8db9f9a