Move const validation out from parsing_driver into ast_validator
Summary:
Instead of validating const values at parse time it instead moves that into the validation step, to be with the other validation logic.

This does result in a slight change in output, in that the previous behaviour was to error out on the line after the offending const, while current behaviour will be to error out on the line, but this seems more correct anyway.

Reviewed By: Alfus, yfeldblum

Differential Revision: D28734654

fbshipit-source-id: 9f8828d2736c02b1c66d1e99188b8d84db8640ce