Fix standard type name generation for typedefs
Summary:
Previously, the 'standard' type name for a typedef was just the typedef's name. This happens to work in most cases because:
 - c++ does not distinguish a typedef from the type itself (i.e. they are interchangable)
 - the type itself happens to be the 'standard' type when the type has not been adapted.

A 'standard' type is a type the Thrift runtime knows how to use. For example, the c++ runtime knows how to handle either std::string or folly::IOBuf for a binary field. The 'cpp.type' annotations is used to change the standard type used.

However, if the type used by the typedef has been adapted, the runtime gets a type it doesn't know how to handle (the adapted type) and the resulting code won't compile. This diff fixes the issue by traversing the typedef instead of just using its name, when generating the standard type name (something it should have been doing anyways).

Unfortunately, using the typedef instead of the real 'standard' type, hide some namespacing issues, so I've only enabled the new path for typedefs with adapters. I will follow up with validators and other changes to address this issue, then apply this change to all cases.

Reviewed By: yfeldblum

Differential Revision: D29683389

fbshipit-source-id: 8a96e92dd22b98b8ace9fd983408419cab6d2fee