Fix cpp.adapter for typedefs and binary fields.
Summary:
Binary fields are special cased in a lot of places; however, this code should not apply to adapted binary types.
Also fixed an issue where the adapter wasn't being found when defined inside a typedef.

At some point we will add support for the adapter to define custom == and < definitions, but for now we just use the ones defined on the adapted type.

Reviewed By: Mizuchi

Differential Revision: D29672670

fbshipit-source-id: fa2cc1c079575288a97b51af39065bf176d217e5