Make Rocket and Upgrade client work in nested EventBase on fibers
Summary: RocketClientChannel can detect that it's running on a fiber and switch to a more efficient mode allocating its state directly on fiber stack. To detect this it validates that sync_ API is used and channel's EventBase is being looped. Unfortunately this heuristic fails if someone creates a separate EventBase just for the client object (since sync_ method implementation loops that EventBase). The fix is to use runInMainContext if EventBase is looped.

Reviewed By: praihan

Differential Revision: D29705420

fbshipit-source-id: 385c0873ebaa25d933b2402d4fdf8e6ef649789c