Handle invalid bools on serialization in compact protocol
Summary:
Detect invalid bools on serialization in the binary protocol. Although we are trying to force values into the valid range via `value ? 1 : 0`, the compiler can and does optimize this away. As a result we get serialized data that cannot be deserialized because we do validation on deserialization side.

This is the same as D27784082 (https://github.com/facebook/fbthrift/commit/e254f78fe24b24f9c1661a6d242c0b0293fe60f4) but for compact protocol. Once we switch from exception to `CHECK` (in a separate diff) `validate_bool` will be reused in both protocols to avoid duplication.

Reviewed By: iahs

Differential Revision: D29682646

fbshipit-source-id: d1ad19a604a6bc7888720aa2bd93beed49d8737b