Add extractPayload to SerializedResponse and LegacySerializedResponse and ResponsePayload::transform
Summary: Add a number of versions of extractPayload methods to the SerializedResponse and LegacySerializedResponse. These take a boolean to indicate whether or not the envelope is required. Normally this will be the result of calling includeEnvelope() on the ResponseChannelRequest. The rest of the parameters vary by the particular scenario but they allow us to migrate code to create SerializedResponse and then only pay the cost of constructing the envelope when requested.

Reviewed By: sethdelliott

Differential Revision: D28976281

fbshipit-source-id: aebac360ff367d8a8fbda648094d936a7a71c708