handle TApplication exception in RSocketServer
Summary:
We follow up from D29796349 where we introduced `ThriftServerRSocket` and handle response and user declared exception correctly.

In this diff, we make sure `TApplicationException` is handled properly as well.
To achieve this, we
* Merge `createPayloadFromTApplicationException(...)` and `createWithTApplicationException(...)` from `GeneratedUtil` into a single method `fromTApplicationException(...)` which takes in `TApplicationException` and returns `ServerResponsePayload` directly.
* Inside this new `fromTApplicationException(...)` method, we initialize `ResponseRpcMetadata` properly with proper `PayloadExceptionMetadataBase`.
* Since similar logic can be shared between generated `RpcServerHandler` and reflection based `RpcServerHandler`, we renamed `GeneratedUtil` -> `RpcPayloadUtil`
* Added a unit test to make sure `TApplicationException` is returned to the client when NPE is thrown from server implementation.

In the coming diffs, I will streamline RSocket server creation so that we can rewrite these unit tests without the usage of `LocalTransport` (similar to what we currently have in `SimpleThriftClientTest`

Differential Revision: D29797624

fbshipit-source-id: f01e998e2cc2fc7cac96970a5537cbc7b9b1e55c