Remove set_x API code generation
Summary:
We had added a setx_ method for thrift unions as a means to provide a setter API that can be used with nullable input.
After some discussions, we decided to revert this method as it seems to be misleading and not the kind of API that we want to support in the long run.

Codemod removed all the uses of this API in WWW.

Reviewed By: vitaut

Differential Revision: D30075621

fbshipit-source-id: 24cf685a5eda901825efde634d0bb17806e3b922