Update creditValSize to use 63 bits even on 32-bit architectures.
Summary:
This fixes the 32-bit ARM build (armv7hl on Fedora).

It turns out that using a :31 bitfield on an uint64_t will still produce an 8-byte field on armv7, even though it produces a 4-byte field on i386.

For consistency with 64-bit platforms, let's always use a 63-bit field for the credits val counter.

Update the assertion to check that the pointer is not larger than the 64-bit counter (which would cause a problem), but allow it to be smaller.

Reviewed By: yfeldblum

Differential Revision: D30050061

fbshipit-source-id: 08bf345fb959aef77ac831d5c42333d46e5df462