simplify lazy deserialization logic
Summary:
This diff is pure refactoring that doesn't change existing logic.

For lazy deserialization, we store last deserialization protocol in `__fbthrift_protocol_`, so that we know which protocol should be used to deserialize lazy IOBuf.

If we deserialize data with new protocol, the existing logic is tracking all fields that's not in serialized data, then deserializing them explicitly.

However, if new protocol is different, we can simplify this logic by just deserializing all fields (instead of tracking which field is already deserialized). Eventually we have to deserialize all fields anyway when protocol changed.

Reviewed By: iahs

Differential Revision: D30327405

fbshipit-source-id: 9558d0b622dd99f5a0d171e588ee8d7219c7f1bb