Remove RocketClientChannel::clientDestroyed_
Summary:
This member is a holdover from when RocketClient was allocated separately. Now that the lifetimes are linked, we aren't at risk of segfaulting when calling methods on RocketClient. Unconditionally call safe methods (since the dead writes don't hurt anything) and guard the reamining few on getTransportWrapper() since socket_ is reset in closeNow().
This member requires pulling in a separate cacheline in several methods. We are already pulling in the socket, so querying it is much cheaper.

Differential Revision: D30342324

fbshipit-source-id: ecebb04b89b99cade52a1c865432b73fe718ebd9