Record OS Thread Ids in ThreadManager and use WorkerProvider to collect them
Summary:
Similar to  D30133823, we are adding the ability for workers in the
ThreadManager to record their OS assigned Thread IDs in a tracking list.
We also add an implementation of `folly::WorkerProvider` to the ThreadManager
for collecting the thread IDS which can be used laster to capture stacks
for dogpiles.

The next diff will contain the implementation of the actual stack collection
in the QueueLagDetectorFIFO observer.

Reviewed By: mshneer

Differential Revision: D30146120

fbshipit-source-id: 6d71fdb9dbd41695c4008f4a41a82f267e2b6342