fix IOBuf union field bug
Summary:
Fixing a bug of setting a union's folly::IOBuf field in constructor. The original implementation `release()` the unique_ptr of a cloned IOBuf, which leaves the cloned IOBuf never get destroyed.

When setting a IOBuf union field, the `IOBuf::IOBuf(const IOBuf& other)` the copy constructor is called and it won't take ownership of the IOBuf instance passed in, thus the unique_ptr shouldn't be released.

Reviewed By: rblaze

Differential Revision: D30390167

fbshipit-source-id: 7ebc9c6dc390e62246457997410d5cc80735cdd4