Explicitly clear the cache on shutdown to avoid memory leaks
Summary: The thrift compiler has memory leaks in the code generators, because the cached entries tend to keep references to the cache, or to a context, which references the cache, and anyway this causes a cycle. Fortunately nobody makes a cycle that includes the generator itself, and so the easy way to break the cycle is to clear the cache explicitly on shutdown.

Reviewed By: yfeldblum

Differential Revision: D30054927

fbshipit-source-id: be88a5eec8f02a3e180f9eb5683b73f2402641aa