Fix flaky ScopedServerInterfaceThread test
Summary: ScopedServerInterfaceThread./[23].joinRequestsStress was failing when run in opt mode due to the server load shedding requests from its Thrift Queue (with Queue Timeout error)  when server is shutting down (see `Cpp2Worker::requestStop`). This shouldn't be considered a failure as the server is functioning as intended. This diff adds an additional exception for queue timeout errors received by the client.

Reviewed By: mshneer

Differential Revision: D30412851

fbshipit-source-id: 15827ded9f15d46dd6a1ac1c1848c67d552d5bb6