Get isset index
Summary:
We want to replace `isset` struct from bool to bits, which can reduce memory usage by compressing 8 bits bool to 1 bit. In order to do this, we need to calculate `isset` index first, then we can set/get `isset` bit base on the index.

So In this diff, we calculate the `isset` index.

- Add `isset_index` function to calculate `isset` index  and memorize it in `isset_index_memo_`
- Register `isset_index` function in `mstch_cpp2_field`
- Move `field_has_isset` to `util.h`, so that it can be reused.
- Add unit test for `isset_index` function.

Reviewed By: Mizuchi

Differential Revision: D30241482

fbshipit-source-id: 230c0329ac9e248e9a5b6746c19ab80927417cc2