Add overloads for RequestChannel::sendRequest* taking rvalue RpcOptions
Summary:
Historically channels used in production did not need a copy of RpcOptions, so the interface took it by &. But now that we are recommending universal use of thread-safe channels like PooledRequestChannel and SR that do need a copy, and we are adding more string fields to the class so that copying is a significant cost, we should revisit this choice.

We could eventually only support the rvalue overloads, but that would require a careful codemod: we currently support reuse of the class between RPCs (though not concurrently, since we write read headers to it). For now we support both, and have them call each other so channels can choose which to implement and users can choose which to call independently.

See stacked diff for perf numbers.

Differential Revision: D30397603

fbshipit-source-id: b7ff25c9a99a833a4760a2ac8a8d3152fb6881da