Add an error handler for cancelled coroutines
Summary:
In python 3.8 coroutine cancellation got moved from `Exception` to `BaseException` which means it needs to be handled separately.

If a coroutine is cancelled we still want the service handler to run normally so I just put a handler to catch the cancellation here. In theory we could just handle all of BaseException but I think things like SIGTERM, SIGINT, etc, shouldn't be handled here.

Also quite embarrassingly it looks like I forgot to fix a TODO in the generic error handler for server streaming, so just rolling that into this diff.

Reviewed By: mshneer

Differential Revision: D30525078

fbshipit-source-id: 4f6a2cee1dc1f4623bb353a5b21966bd8adf7f6f