Fix assignment bug Promise.create method
Summary:
This diff fixes an issue discovered in the generated Cython folly::Promise wrapper's create method.

The issue:
When `create(...)` is called, it first calls `__new__` to create a new instance of the Promise wrapper. Previously, this process created a new (non-empty) instance of `folly::Promise` using the default constructor (since it is a member variable). This default constructor creates a "full" Promise (with a Core for syncing with a corresponding Future). Immediately afterwards, the `folly::Promise` instance passed to `create(...)` is move assigned to `inst.cPromise`.

This change avoids creating a full `folly::Promise` in `create(...)`, which has the following benefits:
1. Doesn't create a Core for the unused placeholder Promise that is subsequently moved into
2. On move assignment, the `detach()` process (including creating a BrokenPromise exception instance) is short circuited

Reviewed By: mshneer

Differential Revision: D30489342

fbshipit-source-id: 81abd39c73b11c74eab2628a1e373166c8c77b45