Add comparison operators to source_loc
Summary: I've used the 'path' to order programs, which seems to me to be the most intuitive approach (though ultimately only stability matters). Equality just checks pointer equality (two programs that have the same relative paths are not necessarily the same 'program'), so I use the address of the instance to break ties and match equality semantics.

Reviewed By: Mizuchi

Differential Revision: D30445504

fbshipit-source-id: f653f61d662041ab2149b79b355a288913af4e9e