table-based thrift-py3-lite serialization/deserialization: part 2 Python changes
Summary:
This is part 2 of the changes: Cython/Python changes
===================
This is the core serialization/deserialization implementation for thrift-py3-lite types:
1. based on cpp2 table based serializer
2. use cpp2 protocols (compact and binary for now)
3. no dependency on generated thrift-cpp2 code
4. generated py code is pure python so no Cython/C++ build system needed for running generated code (the common library is still Cython/C++ but as they are not generated for different thrift IDL it's not an issue).

This implementation uses a Python tuple data structure to represent a Thrift struct dynamically. Python tuple store `PyObject*` (i.e. tuple members) contiguously and uniformly (i.e. `sizeof(PyObject*)` for each element) provides an convenient way to represent any thrift struct and access its fields by a address offset, which is required by table-based serializer.

Each field is internally a `PyObject*` in the tuple. For primitive types, the corresponding Python primitive type is used (e.g. `bool` for `bool`, `int` for `byte`, `i16`, `i32`, `i64`, `str` for `string`, `bytes` for `binary` etc). For `list`, a `tuple` is used, and no surprisingly, `frozenset` for `set`, and `tuple` of `tuple`s for `map` (to support hashing).

A regular Python class is generated for each Thrift struct, and a "spec" is provided as the look-up table for table-based serializer. The metaclass will be in charge of convert that spec into the representation that the c++ table-based serializer may use (i.e. `StructInfo`) for serialization.

In this diff, hand crafted `types.py` files were created to test the implementation.

TODO (in following diffs):

1. Thrift compiler
2. buck macro
3. typing support

Reviewed By: vitaut

Differential Revision: D28264151

fbshipit-source-id: b70b7fba784edf5d949ad402acd410e2f79157e5