Use SerializedResponse in generated code, ResponsePayload in AsyncProcessor
Summary:
Change AsyncProcessor and generated code to create SerializedResponse instead of LegacySerializedResponse and absorb the difference when calling extractPayload() on the response.

This means that when the end of the ResponseChannelRequest chain is rocket we won't ever write the envelope but if its not the extractPayload() method will prepend the envelope header in extractPayload.

Reviewed By: sethdelliott

Differential Revision: D28815348

fbshipit-source-id: d6be6c34f3b69b26df290f4cdcb40e932be94b1f