Add the cpp.ExperimentalAdapter annotation
Summary:
Add the `cpp.ExperimentalAdapter` annotation that applies a C++ adapter to fields. It has the same functionality as `cpp.adapter` but is a structured annotation and works on the field level. For example:

```
struct User {
  cpp.ExperimentalAdapter{name = "IdAdapter"}
  1: i64 id;
}
```

Here the field `id` has the C++ adapter `IdAdapter`.

It is called `ExperimentalAdapter` to discourage use since the exact API is likely to change.

Reviewed By: Alfus

Differential Revision: D30639062

fbshipit-source-id: ae35ce90be1e74b6ebea28a8d88ebbb4b5d93b54