rust/thrift: add on_read_data/on_write_data to the ContextStack trait
Summary:
It seems quite a few Thrift server event handlers key their actions of
`onReadData` so we need to implement it to play well. This diff adds
all the infrastructure, for on_read_data/on_write_data and the next one
actually adds the calls to the generated code.

This diff extends ContextStack to add associated types for `Name` (to match
RequestContext) and `Buffer` (the buffer type being used for the raw framed
messages) so that the `SerializedMessage` struct can have fields matching what
the underlying implementation needs.

(TODO: Do we need the Name assoc type for RequestContext & ContextStack? The
codegen effectively hardcodes this to `const_cstr::ConstCStr`. Is there some
way we can make this more generic?)

Reviewed By: dtolnay

Differential Revision: D30746220

fbshipit-source-id: 69c81d4a4fbf678a40be68bc3caf51ea0e62bd9a