thrift/rust: factor out arg deserialization more like a normal struct
Summary:
Define a structure representing each method's args and implement Deserialize
for it. This makes the server-side method handler code a little easier to read,
easier to put tracing in. Ideally I would have just use the standard struct
serialization generator, but there's enough differences that it wasn't worth
while:
- {{struct:fields}} vs {{function:args}}
- args are always required and there's no need for any optional value handling

In theory there's some risk of extra moves of the args struct, but that should
only be expensive if the args are themselves large structs (not strings, maps,
lists, etc). And by making the function `#[inline]` it should give the compiler
plenty of opportunity to eliminate redundant moves (and the inline itself won't
cost anything because the function is only called from a single site).

I was hoping this would expose opportunities to lift more common code into the
fbthrift crate, but that didn't happen. There might be more opportunity for
dealing with the return values.

#forcetdhashing

Reviewed By: dtolnay

Differential Revision: D30746221

fbshipit-source-id: 0b05e09abb800adf0207bf82ab40321f0030a256