Add ast_mutator
Summary: It turns out templatizing diagnostic_context based on mutability breaks a lot of things, as some stages have mutable access to the AST, and others do not, but they all want to report diagnostics. So instead, I've added a separate mutable visitor context for mutators, which I plan to use to add things to the program or parent nodes in a mutator.

Reviewed By: vitaut

Differential Revision: D30746302

fbshipit-source-id: f49eafd2a0b6a7513fa50d3bbb38328bd0043b55