apply C++ namespaces to pluggable-funcs
Summary: Right now, pluggable-funcs is a private library. But it may be desired to make it a public library. As it is now, though, doing so would present a problem for proper namespacing. It must be possible for multiple libraries to interoperate when each uses pluggable-funcs, even with the same name, so long as each library keeps itself in its own namespace.

Reviewed By: praihan, mshneer, ycking21

Differential Revision: D30265500

fbshipit-source-id: bd7f30685d1e9c294da7b12dd538f7b6d47ee98d