Use value initialization when no custom default is provided for fields that needs to be explicitly constructed
Summary:
Currently `mstch_cpp2_struct::filtered_fields` checks whether a field needs to be explicitly constructed with the criteria of:
1. enum
2. scalar
3. string && not_empty
4. container && not_empty
5. struct && (not_empty || (explicit_ref && not_optional))
6. container && explicit_ref && not_optional

Use value initialization when no custom default is provided for fields that needs to be explicitly constructed. This allows a compile time optimization, such as constexpr default constructor, when the intrinsic default is needed. This is also be beneficial for avoiding type ambiguity from explicit constructor with zero with scalar types.

For example,
```
Foo::Foo() :
      intField(0), // type ambiguity from 0
      optionalIntField(0),
      intFieldWithDefault(13) {
}
```
to
```
Foo::Foo() :
      intField(),
      optionalIntField(),
      intFieldWithDefault(13) { // handled in next diff D30782083
}
```

#forcetdhashing

Reviewed By: Alfus, vitaut

Differential Revision: D30654690

fbshipit-source-id: 5fc097f36292f4e2f7b3af3035d26727cbae9a75