PolledServiceHealth
Summary:
`PolledServiceHealth` exposes a mechanism for a `ServiceHandler` to dynamically communicate to the running `ThriftServer` about the service's health (currently only `OK` and `ERROR`).

The `getServiceHealth` method will be polled periodically by `ThriftServer` and this information will be available to `StatusServerInterface` (set via `ThriftServer::setStatusInterface`).

If the handler needs to work on platforms without coroutines, the user should override `semifuture_getServiceHealth`. Otherwise, `co_getServiceHealth` will work (and will become the default API in the future).

Reviewed By: Gownta

Differential Revision: D31058146

fbshipit-source-id: 3d0837fc472057333eac7b536e298bc335792670