py3lite: generate types for function args
Summary:
For each function, generate an internal struct `_fbthrift_<SERVICE_NAME>_<FUNCTION_NAME>_args`, to be used in clients & servers.

Because both clients and servers need them, putting them in types is the easiest way w/o conflicts.
If import cost becomes a concern in the future, we can extract them out to a separate internal module and let clients & servers depend on it.

Reviewed By: Filip-F

Differential Revision: D31240850

fbshipit-source-id: 8a57fc8adc29d1c97d2f80e0ebf8bdbfbff13171