thrift/rust: refactor non-streaming client reply handling (+split out streaming)
Summary:
The non-streaming client is much simplified by being able to use an async
block.

This also refactors the reply handling code in two ways:
- `fbthrift::help::deserialize_response_envelope` can deserialize an envelope
  either into a Reply Exn type or an ApplicationException.
- a `From<>` implementation which allows a *Exn type (as deserialized off the
  wire as a reply) to be converted to `Result<ReturnType, *Error>`

Together these eliminate a lot of repeated code in reply handling.

In total the test fixtures are a couple of thousand lines smaller (~5% of the
total).

This drops spawning deserialization into a separate task for now,
since I'm not really sure what the value of that is. It should be easy to add
back though.

I tried using `async_trait` to make the whole function async, but it started
requiring more `Sync` constraints. Can look at that later.

#forcetdhashing

Reviewed By: guswynn

Differential Revision: D31164697

fbshipit-source-id: 1fd9cb086391a8a2464565913e2c637056caced8