Swtich to sampled logging of queueing times in responses
Summary:
We added queueing time (process delay) to `ResponseRPCMetadata` for
tracking delays at a per-request level in SR scuba dataset. The value was
only logged if the queueing delay of the request exceeded a pre-configured static
threshold (default of 5 ms). However this would lead to bimodal behavior since
we'd be logging the delay iff the server was already loaded (there is queueing) and not otherwise.
Switching to sampling (with a configurable sampling rate) would make the behavior
identical in either case (server is loaded or not).

(Note: this ignores all push blocking failures!)

Reviewed By: andriigrynenko

Differential Revision: D31309432

fbshipit-source-id: c6ac9c84d17ec3ffb595a1454d8b521468248b7a