Add support for passing object reference and field ID to adapter
Summary:
Add support for passing an object reference of a containing Thrift struct and a field ID to adapter API via an additional argument to `fromThriftField`. For example:

```
struct AdapterWithContext {
  template <typename Context>
  static AdaptedWithContext fromThriftField(int64_t value, Context&& ctx);
  //...
};
```

Here `Context` is `AdapterFieldContext<Struct, FieldID>`.

A Field ID is passed as a template parameter to enable compile-time access and in particular a lookup of static per-field metadata that can be emitted by Thrift plugins.

Reviewed By: Alfus

Differential Revision: D31082742

fbshipit-source-id: 523bf34c02f2fc61180d06e80b99d65717c15a77