Add deterministic protocol tests
Summary:
**This stack:**
The goal of this stack is to introduce a framework/utility to produce hashes in a deterministic way.
The framework will allow users to control the hashing mechanism and will visit thrift structs in a consistent and stable manner.
Users will implement a small class that will dictate how to produce partial hashes for thrift primitive types.
For now code lives in crypto.
Names not final, it's a POC.

**This diff:**
Add protocols test using a debug hasher.
Test that collection ordering does not matter.
Test one big struct with all types.
Test proper handling for optional/empty fields.

Reviewed By: Mizuchi

Differential Revision: D31321225

fbshipit-source-id: 9601c74677152d80a31a08322f5c2b6ffe41ddb3