Pass field adapter context during construction
Summary: The adapter API supports passing an object reference of a containing Thrift struct and a field ID to adapter API via an additional argument to `fromThriftField`. However, this API does't work with constructors because (1) we don't call `fromThrift[Field]` from default constructors and (2) the Thrift object may not be fully constructed at that point. To address this add the optional `construct` method to adapter that can be used to perform the necessary initialization at construct time. If the adapter doesn't provide this method the behavior remains the same.

Reviewed By: thedavekwon

Differential Revision: D31554800

fbshipit-source-id: e5b9df84fc6fd184939d2014f7e193c3a2b2c9a5