Add a validator for custom default integers and floating points
Summary:
Add a validator for custom-defaulted fields with following types:
* byte
* i16
* i32
* float
This validator will check whether a custom default value is within the range of the provided type.

`i64` and `double` can not be validated from validator, and i64 is currently handled in the parser

Currently, it generates a warning, but it will be changed to error when all invalid custom defaults are changed.

Reviewed By: vitaut

Differential Revision: D31426198

fbshipit-source-id: c72a21b034d2dd0d542cb93f611215fee83dd64f