Use SerializedResponse in generated code, ResponsePayload in AsyncProcessor (retry)
Summary:
Change AsyncProcessor and generated code to create SerializedResponse instead of LegacySerializedResponse and absorb the difference when calling extractPayload() on the response.
This means that when the end of the ResponseChannelRequest chain is something that returns false to includeEnvelope() we won't ever write the envelope but if its not the extractPayload() method will prepend the envelope header in extractPayload.

Reviewed By: yfeldblum

Differential Revision: D30852464

fbshipit-source-id: b00157bf59b1ced10829ba6e0e7dc9e7b251d0f5