Fix invalid use of folly::Random in NimbleProtocolTest
Summary: `folly::Random::rand64` only accepts and returns unsigned values, but here it's used with negative values. Interestingly up to 32 bits we were (almost) doing the right thing, and then broke the 64-bit case. I say "almost" because the `max` argument in `rand64()` is exclusive, so it misses one value. So add a helper that always returns the full range for any integral type.

Reviewed By: yfeldblum

Differential Revision: D31641965

fbshipit-source-id: bdde5a6cef3fdefb6c33caace7f272ee609a1d69