disable lazy deserialization on validation failure
Summary:
If for some reasons validation failed, we should disable lazy deserialization to prevent continuous failure and recover as much as we can.

This diff added atomic flag to make it TSAN safe. We use gflags validator to make sure when we update gflags, we will update atomic flags as well.

Reviewed By: yfeldblum

Differential Revision: D31050391

fbshipit-source-id: 603f96a39b240d018af43e1510a050d3b3420fae