support omitting index field - I
Summary:
For the use case of using `cpp.Lazy` for persisted data, we would like
to minimize overhead. As one of the optimizations, let's support omitting index
field if there are no lazy fields present (fields are either optional, or
erased from on-wire representation with `terse_writes`), which should shave off
at least ~5 bytes in common use case (that is 4B for field id in case of
compact protocol [thanks to delta coding of i32 fields ids] + 1B for size).

The idea is to just use NaN as a sentinel to signal about the absence of index,
and this diff wires in read part.

Reviewed By: luciang

Differential Revision: D31788071

fbshipit-source-id: 274cd8f8d363efd59b780635fa796d68de74a142