remove _ref() suffix (3): mangle data member name when eligible
Summary:
If field is eligible, this diff change data member name to "__fbthrift_field_{name}_", so that `name` can be reused for field accessor.

Currently there are some corner cases that are not handled, here is the list of cases where we don't mangle data member name

1. tablebased serialization or reflection is enabled
2. field is cpp.ref or field has adapter
3. field is not private
4. struct has `cpp.methods`

For cases (1) and (2), there is nothing stops us from mangling the name. It's just that this is not implemented yet. We can cover most case first, then handle these cases later.

Reviewed By: vitaut

Differential Revision: D31614483

fbshipit-source-id: 35727d9ae26e618002b079e4f6d287c9b6444753